requirements.txt
langchain==0.1.7
python-telegram-bot==20.8
schedule==1.2.1
openai-whisper==20231117